pyproject.toml
[project]
name = "agentic-trading-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.2.0",
    "massive>=2.0.2",
    "mcp>=1.25.0",
    "openai>=2.14.0",
    "openai-agents>=0.6.4",
    "playwright>=1.57.0",
    "plotly>=6.5.0",
    "polygon>=1.2.8",
    "python-dotenv>=1.2.1",
]
